More correctly state what happens when iterating over Rust array types
As of Rust 1.48, array types do not actually implement IntoIterator, which
means that the borrow checker automatically introduces borrows with
surprising results. This behavior is now explicitly marked as unstable:

warning: this method call currently resolves to
`<&[T; N] as IntoIterator>::into_iter` (due to autoref coercions), but
that might change in the future when `IntoIterator` impls for arrays
are added.
warning: this was previously accepted by the compiler but is being
phased out; it will become a hard error in a future release!
note: for more information, see issue #66145
<https://github.com/rust-lang/rust/issues/66145>

This commit changes the into_iter() calls into iter() calls that
match the current behavior in Rust.